feat: EmptyComponent 구현 및 적용
