+++ src/Page/Cart/index.module.scss
@@ -6,30 +6,6 @@
   column-gap: 32px;
 }
 
-.main-view-blank {
-  display: flex;
-  flex-direction: column;
-  row-gap: 24px;
-
-  align-items: center;
-  width: 100%;
-  justify-content: center;
-
-  & > p {
-    font-weight: 700;
-    font-size: 28px;
-  }
-
-  & > button {
-    width: 200px;
-    height: 40px;
-
-    background: var(--mint-color);
-    color: white;
-    border-radius: 6px;
-  }
-}
-
 @media screen and (max-width: 1024px) {
   .main-view {
     display: flex;
